Downloads/project 5/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
mysql-connector-python==8.2.0
python-multipart==0.0.6
pydantic==2.5.0